+++ tests/test_event_reader.py
@@ -157,11 +157,8 @@ def test_read_events_concurrent_two_blocks_concurrent(web3):
     assert max(blocks) == 37898276
 
 
-
 def test_read_events_lazy_timestamp(web3):
-    """Read events but extract timestamps only for events, not whole block ranges.
-
-    """
+    """Read events but extract timestamps only for events, not whole block ranges."""
 
     # Get contracts
     Pair = get_contract(web3, "sushi/UniswapV2Pair.json")