+++ CHANGELOG.md
@@ -3,6 +3,20 @@
 - Adds support for extracting block data using Ankr
 - Adds `AnkrReorganizationMonitor`
 
+# 0.22
+
+- Refactor a lot of functionality to a new submodule [eth_defi.provider](https://web3-ethereum-defi.readthedocs.io/api/provider/index.html) 
+- Add MEV blocking support in the form of `eth_defi.mev_blocker.MEVBlockerProvider`
+- Add JSON-RPC fallback switching in the form of `eth_defi.fallback_provider.FallbackProvider`
+- Add `HotWallet.create_for_testing` 
+- Add utility function `get_onchain_price()` to ask on-chain price of a 
+  Uniswap v3 pool at any given block number
+- Add `eth_defi.event_reader.logresult.decode_log` and better 
+  documentation for `LogResult` class
+- Deprecate `eth_defi.anvil` -> `eth_defi.provider.anvil` 
+- Deprecate `eth_defi.ganache` -> `eth_defi.provider.ganache`
+
+
 # 0.21.8
 
 - Add test coverage for `extract_timestamps_json_rpc_lazy`